Combine the utility classes into the helper
This fixes the issue where the select sometimes throws an uninitialized
constant - https://github.com/GetJobber/recurring_select/pull/22. I
have combined the require_relative classes into the main helper class,
which means that autoload doesn't break with rails4.
